refactor: move single-component integration tests nearer to their target
